refactor: rename input validation
